complex + тесты

+++ lib.h
@@ -1,8 +1,8 @@
 #pragma once
 
 struct complex {
-    double real; // действительная часть
-    double imaginary; // мнимая часть
+    float real; // действительная часть
+    float imaginary; // мнимая часть
 };
 
 complex operator+ (complex a, complex b);
@@ -16,5 +16,6 @@ void operator/= (complex &dividend, complex divisor);
 
 complex operator* (complex multiplier, complex multiplicand);
 void operator*= (complex &multiplier, complex multiplicand);
+
 
 std::ostream& operator<< (std::ostream& out, complex a);
